Pams refined/requirements.txt
customtkinter>=5.2.0
pymongo>=4.6.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
matplotlib>=3.8.0
Pillow>=10.0.0
dnspython>=2.4.0
